feat(engine)!: get / get_by_id は列の幅によらず u64 を返す (get64 / get_by_id64 を吸収)
u32 に収まらない 64 bit 列の値を `get` が None で返していた (値が無いのと区別できない)
のをやめ、 `get` / `get_by_id` を `Option<u64>` にした。 `get64` / `get_by_id64` は消える。

- engine 内部の u32 の列 (_sync_ops の lsn / vocab id / ref) は `get_by_id32` (crate 内、
  64 bit 列で panic) で読む
- Ravn の `path` / `extract` / `select` / `RavnResult::Values` の値も u64
- enchudb-rag の `meta_value` も u64
- 下流の破壊的変更: `get` の結果を u32 として使う所は、 Ref なら
  `make_eid(peer, v as u32)`、 数値なら u64 のまま比べる

+++ crates/enchudb-engine/tests/engine_model_proptest.rs
@@ -145,7 +145,7 @@ fn verify(
                 Kind::Num => {
                     let got = eng.get(eid, hname);
                     let want = exp.map(|v| match v {
-                        Val::Num(n) => *n,
+                        Val::Num(n) => u64::from(*n),
                         Val::Txt(_) => unreachable!(),
                     });
                     prop_assert_eq!(got, want, "get(e{}, {}) mismatch", e, hname);
